+++ src/main/java/com/itdr/pojo/vo/CartProductVO.java
@@ -28,7 +28,7 @@ public class CartProductVO {
 
     private Integer price;
 
-    private BigDecimal sumprice;
+    private Integer sumprice;
 
     private String pimages;
