+++ src/main/java/com/itdr/service/AliPayService.java
@@ -10,9 +10,9 @@ import java.util.Map;
  * @date 2020/2/26 18:43
  */
 public interface AliPayService {
-    ServerResponse pay(ItdrUser user, Long orderNo);
+    ServerResponse pay(ItdrUser user, Integer orderNo);
 
     ServerResponse alipayCallback(Map<String, String> params);
 
-    ServerResponse queryOrderPayStatus(ItdrUser user, Long orderNo);
+    ServerResponse queryOrderPayStatus(ItdrUser user, Integer orderNo);
 }
